requirements.txt
asgiref==3.6.0
attrs==22.2.0
cffi==1.15.1
coverage==7.0.1
cryptography==39.0.0
Django==4.1.5
djangorestframework==3.14.0
exceptiongroup==1.1.0
Faker==12.0.1
iniconfig==1.1.1
jwt==1.3.1
mixer==7.2.2
packaging==22.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
